Bugfixes. Changed L-EST97 coordinates to be read in as BigDecimal to ensure integral precision on conversion and avoid potentially missing blocks.

+++ src/main/kotlin/org/mcraster/converters/PointLEst97.kt
@@ -1,10 +1,12 @@
 package org.mcraster.converters
 
+import java.math.BigDecimal
+
 /**
  * X and Y stand for coordinates from the L-EST97 system, H stands for height from sea level (in meters).
  * In the L-EST97 coordinate system, X increases towards North, Y increases towards East.
  */
-class PointLEst97(val x: Double, val y: Double, val h: Double) {
+class PointLEst97(val x: BigDecimal, val y: BigDecimal, val h: BigDecimal) {
 
     fun getBoundingBlock() = BlockPosLEst97.fromPointOnBlock(x = x, y = y, h = h)
